# openapi/finance-api.yaml
openapi: 3.0.3
info:
  title: Personal Finance Management API
  description: A comprehensive REST API for managing personal finances including transactions, budgets, recurring transactions, and analytics
  version: 1.0.0
  contact:
    name: API Support
    email: support@finance.app
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.finance.app/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /categories:
    get:
      tags:
        - Transactions
      summary: Get all categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
    post:
      tags:
        - Transactions
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /transactions:
    get:
      tags:
        - Transactions
      summary: Get all transactions with filters
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: category_id
          in: query
          schema:
            type: integer
        - name: wallet_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [EXPENSE, INCOME, TRANSFER]
        - name: min_amount
          in: query
          description: Minimum transaction amount for filtering
          schema:
            type: number
            format: float
            minimum: 0
        - name: max_amount
          in: query
          description: Maximum transaction amount for filtering
          schema:
            type: number
            format: float
            minimum: 0
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Transactions
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{transaction_id}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Transactions
      summary: Update transaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateRequest'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Transactions
      summary: Delete transaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Transaction deleted
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/daily:
    get:
      tags:
        - Transactions
      summary: Get transactions grouped by day
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions grouped by day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyTransactionGroup'
  /transactions/weekly:
    get:
      tags:
        - Transactions
      summary: Get transactions grouped by week
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions grouped by week
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyTransactionGroup'
  /transactions/monthly:
    get:
      tags:
        - Transactions
      summary: Get transactions grouped by month
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions grouped by month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonthlyTransactionGroup'
...
  /recurring-transactions:
    get:
      tags:
        - RecurringTransactions
      summary: Get all recurring transactions
      responses:
        '200':
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringTransactionResponse'
    post:
      tags:
        - RecurringTransactions
      summary: Create a new recurring transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringTransactionCreateRequest'
      responses:
        '201':
          description: Recurring transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringTransactionResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recurring-transactions/{recurring_id}:
    get:
      tags:
        - RecurringTransactions
      summary: Get recurring transaction by ID
      parameters:
        - name: recurring_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Recurring transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringTransactionResponse'
        '404':
          description: Recurring transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - RecurringTransactions
      summary: Update recurring transaction
      parameters:
        - name: recurring_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringTransactionUpdateRequest'
      responses:
        '200':
          description: Recurring transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringTransactionResponse'
        '404':
          description: Recurring transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - RecurringTransactions
      summary: Delete recurring transaction
      parameters:
        - name: recurring_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Recurring transaction deleted
        '404':
          description: Recurring transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recurring-transactions/{recurring_id}/execute:
    post:
      tags:
        - RecurringTransactions
      summary: Manually execute a recurring transaction
      description: Creates a new transaction based on the recurring template
      parameters:
        - name: recurring_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Transaction created from recurring template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Recurring transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
...
    WalletResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        currency:
          type: string
        balance:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - name
        - currency
        - balance
        - created_at
    CategoryCreateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [EXPENSE, INCOME]
        icon:
          type: string
        color:
          type: string
      required:
        - name
        - type
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [EXPENSE, INCOME]
        icon:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - type
    TransactionCreateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
          default: VND
        date:
          type: string
          format: date-time
        category_id:
          type: integer
        wallet_id:
          type: integer
        to_wallet_id:
          type: integer
          description: Required for TRANSFER type
        description:
          type: string
      required:
        - type
        - amount
        - date
    TransactionUpdateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
        date:
          type: string
          format: date-time
        category_id:
          type: integer
        wallet_id:
          type: integer
        to_wallet_id:
          type: integer
        description:
          type: string
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
        date:
          type: string
          format: date-time
        category_id:
          type: integer
        category:
          $ref: '#/components/schemas/CategoryResponse'
        wallet_id:
          type: integer
        wallet:
          $ref: '#/components/schemas/WalletResponse'
        to_wallet_id:
          type: integer
        to_wallet:
          $ref: '#/components/schemas/WalletResponse'
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - type
        - amount
        - currency
        - date
        - created_at
    TransactionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
      required:
        - items
        - total
        - page
        - size
        - pages
    DailyTransactionGroup:
      type: object
      properties:
        date:
          type: string
          format: date
        total_income:
          type: number
          format: double
        total_expense:
          type: number
          format: double
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
    WeeklyTransactionGroup:
      type: object
      properties:
        year:
          type: integer
        week:
          type: integer
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        total_income:
          type: number
          format: double
        total_expense:
          type: number
          format: double
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
    MonthlyTransactionGroup:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        total_income:
          type: number
          format: double
        total_expense:
          type: number
          format: double
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
...
    RecurringTransactionCreateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
          default: VND
        category_id:
          type: integer
        wallet_id:
          type: integer
        to_wallet_id:
          type: integer
        description:
          type: string
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        start_date:
          type: string
          format: date
        is_active:
          type: boolean
          default: true
      required:
        - type
        - amount
        - frequency
        - start_date
    RecurringTransactionUpdateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
        category_id:
          type: integer
        wallet_id:
          type: integer
        to_wallet_id:
          type: integer
        description:
          type: string
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        start_date:
          type: string
          format: date
        is_active:
          type: boolean
    RecurringTransactionResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
        amount:
          type: number
          format: double
        currency:
          type: string
        category_id:
          type: integer
        category:
          $ref: '#/components/schemas/CategoryResponse'
        wallet_id:
          type: integer
        wallet:
          $ref: '#/components/schemas/WalletResponse'
        to_wallet_id:
          type: integer
        to_wallet:
          $ref: '#/components/schemas/WalletResponse'
        description:
          type: string
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        start_date:
          type: string
          format: date
        next_run_date:
          type: string
          format: date
        last_run_date:
          type: string
          format: date
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - type
        - amount
        - currency
        - frequency
        - start_date
        - next_run_date
        - is_active
        - created_at
...